+++ elide-datastore/elide-datastore-jpa/src/main/java/com/yahoo/elide/datastores/jpa/transaction/NonJtaTransaction.java
@@ -1,0 +1,58 @@
+/*
+ * Copyright 2018, Oath Inc.
+ * Licensed under the Apache License, Version 2.0
+ * See LICENSE file in project root for terms.
+ */
+package com.yahoo.elide.datastores.jpa.transaction;
+
+import com.yahoo.elide.core.RequestScope;
+
+import lombok.extern.slf4j.Slf4j;
+
+import javax.persistence.EntityManager;
+import javax.persistence.EntityTransaction;
+
+/**
+ * Non-JTA transaction implementation.
+ */
+@Slf4j
+public class NonJtaTransaction extends AbstractJpaTransaction {
+    private final EntityTransaction transaction;
+
+    public NonJtaTransaction(EntityManager entityManager) {
+        super(entityManager);
+        this.transaction = entityManager.getTransaction();
+        entityManager.clear();
+    }
+
+    @Override
+    public void begin() {
+        if (!transaction.isActive()) {
+            transaction.begin();
+        }
+    }
+
+    @Override
+    public void commit(RequestScope scope) {
+        if (transaction.isActive()) {
+            super.commit(scope);
+            transaction.commit();
+        }
+    }
+
+    @Override
+    public void rollback() {
+        if (transaction.isActive()) {
+            try {
+                super.rollback();
+            } finally {
+                transaction.rollback();
+            }
+        }
+    }
+
+    @Override
+    public boolean isOpen() {
+        return transaction.isActive();
+    }
+}