+++ elide-datastore/elide-datastore-jpa/src/test/java/com/yahoo/elide/datastores/jpa/JpaDataStoreSupplier.java
@@ -3,25 +3,22 @@
  * Licensed under the Apache License, Version 2.0
  * See LICENSE file in project root for terms.
  */
-package com.yahoo.elide.datastores.hibernate5;
+package com.yahoo.elide.datastores.jpa;
 
 import com.yahoo.elide.core.DataStore;
+import com.yahoo.elide.datastores.jpa.transaction.NonJtaTransaction;
 import com.yahoo.elide.models.generics.Manager;
 import com.yahoo.elide.models.triggers.Invoice;
 import com.yahoo.elide.utils.ClassScanner;
 
-import example.Filtered;
 import example.Parent;
-import example.TestCheckMappings;
 
 import org.hibernate.MappingException;
-import org.hibernate.ScrollMode;
 import org.hibernate.boot.MetadataSources;
 import org.hibernate.boot.registry.StandardServiceRegistryBuilder;
 import org.hibernate.boot.spi.MetadataImplementor;
 import org.hibernate.cfg.Environment;
 import org.hibernate.jpa.AvailableSettings;
-import org.hibernate.jpa.HibernateEntityManager;
 import org.hibernate.tool.hbm2ddl.SchemaExport;
 import org.hibernate.tool.schema.TargetType;
 
@@ -32,25 +29,21 @@ import java.util.Map;
 import java.util.function.Supplier;
 
 import javax.persistence.Entity;
+import javax.persistence.EntityManager;
 import javax.persistence.EntityManagerFactory;
 import javax.persistence.Persistence;
 
 /**
  * Supplier of Hibernate 5 Data Store.
  */
-public class HibernateEntityManagerDataStoreSupplier implements Supplier<DataStore> {
+public class JpaDataStoreSupplier implements Supplier<DataStore> {
     private static final String JDBC = "jdbc:h2:mem:root;IGNORECASE=TRUE";
     private static final String ROOT = "root";
 
     @Override
     public DataStore get() {
-        // Add additional checks to our static check mappings map.
-        // NOTE: This is a bit hacky. We need to do a major overhaul on our test architecture
-        TestCheckMappings.MAPPINGS.put("filterCheck", Filtered.FilterCheck.class);
-        TestCheckMappings.MAPPINGS.put("filterCheck3", Filtered.FilterCheck3.class);
-
         Map<String, Object> options = new HashMap<>();
-        ArrayList<Class> bindClasses = new ArrayList<>();
+        ArrayList<Class<?>> bindClasses = new ArrayList<>();
 
         try {
             bindClasses.addAll(ClassScanner.getAnnotatedClasses(Parent.class.getPackage(), Entity.class));
@@ -68,7 +61,7 @@ public class HibernateEntityManagerDataStoreSupplier implements Supplier<DataSto
         options.put(AvailableSettings.LOADED_CLASSES, bindClasses);
 
         EntityManagerFactory emf = Persistence.createEntityManagerFactory("elide-tests", options);
-        HibernateEntityManager em = (HibernateEntityManager) emf.createEntityManager();
+        EntityManager em = emf.createEntityManager();
 
         // method to force class initialization
         MetadataSources metadataSources = new MetadataSources(
@@ -83,6 +76,10 @@ public class HibernateEntityManagerDataStoreSupplier implements Supplier<DataSto
 
         try {
             ClassScanner.getAnnotatedClasses(Parent.class.getPackage(), Entity.class)
+                    .forEach(metadataSources::addAnnotatedClass);
+            ClassScanner.getAnnotatedClasses(Manager.class.getPackage(), Entity.class)
+                    .forEach(metadataSources::addAnnotatedClass);
+            ClassScanner.getAnnotatedClasses(Invoice.class.getPackage(), Entity.class)
                     .forEach(metadataSources::addAnnotatedClass);
         } catch (MappingException e) {
             throw new IllegalStateException(e);
@@ -100,9 +97,9 @@ public class HibernateEntityManagerDataStoreSupplier implements Supplier<DataSto
             throw new IllegalStateException(schemaExport.getExceptions().toString());
         }
 
-        return new AbstractHibernateStore.Builder(em)
-                .withScrollEnabled(true)
-                .withScrollMode(ScrollMode.FORWARD_ONLY)
-                .build();
+        return new JpaDataStore(
+                () -> { return em; },
+                (entityManager) -> { return new NonJtaTransaction(entityManager); }
+        );
     }
 }