+++ elide-integration-tests/src/test/groovy/com/yahoo/elide/tests/ShareableIT.groovy
@@ -29,6 +29,7 @@ class ShareableIT extends AbstractIntegrationTestInitializer {
         Left left = new Left();
         tx.createObject(left, null);
         tx.commit(null);
+        tx.close();
     }
 
     @Test