+++ elide-integration-tests/src/test/java/example/NoUpdateEntity.java
@@ -23,6 +23,6 @@ import javax.persistence.Table;
 @Entity
 @Table(name = "noupdate")
 public class NoUpdateEntity extends BaseId {
-    @OneToMany()
-    public Set<Child> children;
+    @OneToMany
+    protected Set<Child> children;
 }