+++ elide-core/src/main/java/com/yahoo/elide/core/DataStoreTransaction.java
@@ -142,7 +142,10 @@ public interface DataStoreTransaction extends Closeable {
                 scope);
         Iterator<Object> it = results == null ? null : results.iterator();
         if (it != null && it.hasNext()) {
-            return it.next();
+            Object obj = it.next();
+            if (!it.hasNext()) {
+              return obj;
+            }
         }
         return null;
     }