+++ changelog.md
@@ -6,6 +6,7 @@
 **Fixes**
  * Throw proper exception on invalid PersistentResource where id=null
  * Issue#744 Elide returns wrong date parsing format in 400 error for non-default DateFormats
+ * Enable RSQL filter dialect by default (in addition to the default filter dialect).
 
 **Features**
  * Added [JPA Data Store](https://github.com/yahoo/elide/pull/747)